Add BinomialLeastSquares fitting plugin
This differs from GaussianLeastSquares in the variance calculation step.
It gets rid of the problematic behavior when a (toy) fit has 0 counts
in one of the boxes.
As there are more plugins foreeen to come in upcoming commits,
the default behavior is not yet changed
(even though BinomialLeastSquares seems favorable in all cases).