Fitting: Simplify masking ToyValues
It is possible that the fit (on the expected counts) works flawlessly,
while some of the toy fits have problems.
A covariance matrix that is not accurate is often
a symptom of this problem.
In these cases, derived quantities (e.g. parameter correlations
from the toy population) can be heavily distorted.
One example for this problem are low-expectancy bins fluctuating
to 0 counts for a Gaussian Least Square likelihod.

It can be interesting to derive quantities expluding these toys.
This commit introduces a convenience function for this.
Also, it adds some information about this option.