Consistency: Allways use Minuit.LIKELIHOOD
It is adviced in iminuit.Minuit to use Minuit.LEAST_SQUARES (which =1)
when working with χ² instead of Likelihoods.
For consistency, we will use Minuit.Likelihood (which = 0.5).
This is achieved by plugin a factor 0.5 to all χ² functions,
effectively making them likelihood functions.
By having this consistent notation, it is possible to e.g.
set Minuit.errordef = 2 (-> ×4) for 95% / 2σ errors
in place of the standard 67% / 1σ errors.

Note that this differs from the Minuit default,  which is errordef = 1.
Add a test to ensure that new plugins to not forget to change this.